requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
langchain-community
langchain-google-genai
pydantic-settings
sqlalchemy
python-dotenv
tabulate
psycopg2-binary
